+++ bld/sdk/wde/c/wdeppmem.c
@@ -48,7 +48,7 @@ void PP_OutOfMemory( void )
     }
 }
 
-void *PP_Malloc( unsigned size )
+void *PP_Malloc( size_t size )
 {
     void        *p;
